refactor: Main 페이지 컴포넌트 분리 Detail label fontSize 입력받도록 수정
