+++ client/src/components/Card/RecruitCard/RecruitCard.tsx
@@ -2,7 +2,7 @@ import { Recruit } from "#types/Recruit";
 import { getDisplayPaceString } from "#utils/stringUtils";
 import Card from "../Card";
 import { CLOCK_ICON, LOCATION_ICON, RULER_ICON, RUNNING_ICON, USER_ICON } from "#assets/icons";
-import { CardTitle, PeopleWrapper, SummaryBody, SummaryHead, TimeWrapper, UserIdLabel } from "../Card.styles";
+import { CardTitle, PeopleWrapper, SummaryBody, SummaryHead, UserIdLabel } from "../Card.styles";
 import { timeDifference } from "#utils/cardUtils";
 
 interface RecruitCardProps {
