+++ client/src/pages/Main/MainPage.tsx
@@ -1,57 +1,26 @@
 import Header from "#components/Header/Header";
-import MoreButton from "#components/MoreButton/MoreButton";
-import Slider, { Settings } from "react-slick";
-import CourseCard from "#components/Card/CourseCard/CourseCard";
-import { CarouselWrapper, ListTitle, MainPageContainer, TitleWrapper } from "./MainPage.styles";
-import RecruitTextCard from "#components/Card/RecruitTextCard/RecruitTextCard";
-import useCoursesQuery from "#hooks/queries/useCoursesQuery";
-import useRecruitsQuery from "#hooks/queries/useRecruitsQuery";
-import { Course } from "#types/Course";
-import { Recruit } from "#types/Recruit";
+import RecommendCourses from "#components/RecommendCourses/RecommendCourses";
+import RecommendRecruits from "#components/RecommendRecruits/RecommendRecruits";
+import styled from "styled-components";
 
-const settings: Settings = {
-    centerMode: true,
-    infinite: true,
-    slidesToShow: 1,
-    dots: true,
-    slidesToScroll: 1,
-};
+const MainPageContainer = styled.div`
+    overflow-x: hidden;
+    overflow-y: hidden;
+    height: 100vh;
+    > div:first-child {
+        margin-bottom: 16px;
+    }
+    > div:last-child {
+        margin-top: 48px;
+    }
+`;
 
 const MainPage = () => {
-    const { data: course, isLoading: coursesLoading } = useCoursesQuery();
-    const { data: recruit, isLoading: recruitsLoading } = useRecruitsQuery();
-    if (coursesLoading || recruitsLoading) return <div>Loading...</div>;
-    if (!course || !recruit) return <div>404</div>;
-
     return (
         <MainPageContainer>
             <Header isMain={true} text="RunWithMe" />
-            <div>
-                <TitleWrapper>
-                    <ListTitle>코스 목록</ListTitle>
-                    <MoreButton to="/courses" />
-                </TitleWrapper>
-                <CarouselWrapper>
-                    <Slider {...settings}>
-                        {course.data.map((c: Course, idx: number) => (
-                            <CourseCard data={c} key={idx} />
-                        ))}
-                    </Slider>
-                </CarouselWrapper>
-            </div>
-            <div>
-                <TitleWrapper>
-                    <ListTitle>모집 목록</ListTitle>
-                    <MoreButton to="recruits" />
-                </TitleWrapper>
-                <CarouselWrapper>
-                    <Slider {...settings}>
-                        {recruit.data.map((r: Recruit, idx: number) => (
-                            <RecruitTextCard data={r} key={idx} />
-                        ))}
-                    </Slider>
-                </CarouselWrapper>
-            </div>
+            <RecommendCourses />
+            <RecommendRecruits />
         </MainPageContainer>
     );
 };
